Move sidebar rendering to it's own component
